+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **reset-scoped:** output chunk file names
+
 ## [3.2.1] - 2023-06-05
 
 ### Fixed